refactoring(searchBarAutocomplete): extract useOptions
