+++ datahub-web-react/src/app/searchV2/searchBarV2/hooks/useOptions.ts
@@ -1,0 +1,57 @@
+import { Entity } from '@src/types.generated';
+import { useMemo } from 'react';
+import useSearchResultsOptions from './useAutocompleteSuggestionsOptions';
+import useRecentlySearchedQueriesOptions from './useRecentlySearchedQueriesOptions';
+import useRecentlyViewedEntitiesOptions from './useRecentlyViewedEntitiesOptions';
+import useViewAllResultsOptions from './useViewAllResultsOptions';
+
+export default function useOptions(
+    searchQuery: string,
+    showViewAllResults: boolean | undefined,
+    entities: Entity[] | undefined,
+    isDataLoading: boolean,
+    isDataInitialized: boolean,
+    shouldCombineSiblings: boolean,
+    isSearching: boolean,
+    shouldShowAutoCompleteResults: boolean,
+) {
+    const recentlySearchedQueriesOptions = useRecentlySearchedQueriesOptions();
+    const recentlyViewedEntitiesOptions = useRecentlyViewedEntitiesOptions();
+
+    const initialOptions = useMemo(() => {
+        return [...recentlyViewedEntitiesOptions, ...recentlySearchedQueriesOptions];
+    }, [recentlyViewedEntitiesOptions, recentlySearchedQueriesOptions]);
+
+    const viewAllResultsOptions = useViewAllResultsOptions(searchQuery, showViewAllResults);
+
+    const hasResults = useMemo(() => (entities?.length ?? 0) > 0, [entities?.length]);
+
+    const searchResultsOptions = useSearchResultsOptions(
+        entities,
+        searchQuery,
+        isDataLoading,
+        isDataInitialized,
+        shouldCombineSiblings,
+    );
+
+    const options = useMemo(() => {
+        if (!isSearching) return initialOptions;
+
+        if (shouldShowAutoCompleteResults) {
+            if (!isDataLoading && !hasResults) return [];
+            return [...viewAllResultsOptions, ...searchResultsOptions];
+        }
+
+        return [];
+    }, [
+        isSearching,
+        hasResults,
+        initialOptions,
+        searchResultsOptions,
+        viewAllResultsOptions,
+        shouldShowAutoCompleteResults,
+        isDataLoading,
+    ]);
+
+    return options;
+}
